+++ saleor/static/dashboard-next/components/AddressFormatter/AddressFormatter.tsx
@@ -1,7 +1,7 @@
 import Typography from "@material-ui/core/Typography";
 import * as React from "react";
 
-import { AddressType } from "../../customers/";
+import { AddressType } from "../../customers/types";
 import Skeleton from "../Skeleton";
 
 interface AddressFormatterProps {
